Moved the form errors calculation to a selector
