+++ client/lib/initialize-state/index.js
@@ -27,6 +27,7 @@ export default ( schema, values ) => {
 	return {
 		settings: formValues,
 		form: {
+			schema,
 			isSaving: false,
 			pristine: true,
 			packages: {
