+++ package.json
@@ -81,6 +81,7 @@
     "react-dom": "0.14.7",
     "react-redux": "^4.4.1",
     "redux": "^3.3.1",
+    "reselect": "^2.5.1",
     "sprintf-js": "1.0.3",
     "wp-calypso": "github:automattic/wp-calypso"
   }
